Move the draw packages together
