--- src/data/doc.go
@@ -1,4 +1,0 @@
-/*
- Package data provides structures to store arrays in a hardware-agnostic (GPU-CPU) way.
-*/
-package data
