+++ src/draw/raster/geom.go
