+++ src/draw/raster/paint.go
