+++ src/draw/raster/raster.go
