+++ src/draw/raster/stroke.go
